Refactor CredentialsDataGrid
